editied github actions script and added styling to bookshelf
